backend/requirements.txt
# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# 数据库
sqlalchemy==2.0.23
alembic==1.12.1
sqlite3

# AI和机器学习
torch==2.1.1
torchvision==0.16.1
transformers==4.35.2
flagembedding==1.2.5
sentence-transformers==2.2.2

# 搜索和索引
faiss-cpu==1.7.4
whoosh==2.7.4
jieba==0.42.1

# 文件处理
python-docx==1.1.0
PyPDF2==3.0.1
openpyxl==3.1.2
python-pptx==0.6.23
pillow==10.1.0
opencv-python==4.8.1.78

# 音频处理
openai-whisper==20231117
pydub==0.25.1

# OCR和图像处理
paddleocr==2.7.0.10
chinese-clip-vit-base-patch16==1.5.0

# 语音识别
speechrecognition==3.10.0

# 数据处理
pandas==2.1.3
numpy==1.25.2
scipy==1.11.4

# 文件监控
watchdog==3.0.0

# HTTP客户端
httpx==0.25.2
requests==2.31.0

# 配置和环境
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# 日志和监控
loguru==0.7.2
rich==13.7.0

# 开发工具
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
mypy==1.7.1
ruff==0.1.6

# 安全
cryptography==41.0.7
passlib[bcrypt]==1.7.4

# 工具库
click==8.1.7
typer==0.9.0
tqdm==4.66.1
aiofiles==23.2.1

# 时间处理
python-dateutil==2.8.2
pytz==2023.3

# 类型提示
types-requests==2.31.0.10
types-python-dateutil==2.8.19.14